First step in refactoring how rules are written
You can now write per file and per line rules as classes, which means
they are easier to document, less linked with traversal, and all the
common stuff like checking it applies is handles centrally.

In the future we can extract the docstring to auto-generate a part of
the style checker manual.