Link statement terminators and end keywords to AST
This required a bit of a re-work on how nodes are built. Now more nodes
are built in steps, instead of providing all children at construction
time. We do this to be able to attach punctuation we don't keep in the
AST to the relevant nodes they would belong to.

We also can't generate Simple_For_Statement nodes anymore, so we just
build General_For_Statements. The reason for this is that by the time
we know which one we have, we've already assigned the various punctuation
tokens to the AST. In the future we can always re-write these.