Add exploratory CFG/PDG for a assignment statement
It may be worth to explode out assignments (and anything else) like
this. This is an option we never considered in SPARK flow analysis,
but there were some issues (in particular surrounding calls) where
this would have been useful.

A side benefit of this is that computing cyclomatic complexity no
longer requires a special case for short-circuits.