#139 Mis-classification of & and | inside expressions
While if/while guards change the semantics of & and |, inside
e.g. function calls this change does not happen anymore.

This caused mlint to produce a very different metric for such
code. Now we produce the same metric again.

In addition I disabled the corresponding style rule again, since
it is not correct in all cases. Beware: mlint shares the same
bug and mis-identifies & in the case where the operands are
arrays.