Fix duplicate messages for embedded matlab
The Message_Handler was not forked for each embedded matlab work
package; which meant that integrate was called multiple times (once
for each matlab function), accumulating a message raised each time
we analyse another function.

There is a deeper design problem here, since the blockname was kinda
tacked on and never really part of the original design. To be reworked
in the future.

Also add some code helpful for debugging and comments.

Fixes #161