#149 Add/clarify private tests
Private directories are quite restricted:
* You can only have functions defined
* Not classes
* And also no further packages

So, packages and classes always have global visibility.

Visibility for functions in private directories are also strict:
only items specifically in that directory have visibility.

Items that are logically there (like functions in identically
named packages, or classes in @ directories) do not, since they
are not literally in that directory.