#238 Fix functions named end in classes
Deal with a special issue of having a function named end inside
a class. The lexer misclassified the end (the function name) to
end the function, meaning the end (of the function) then ended the
special block rules for methods, and the next end (of the methods block)
ended the class.

This then made methods not a keyword anymore, causing a super
confusing parse error (as methods was seen as an identifier).

This is now fixed; by a simple hack: we now only consider standalone
end statements to end blocks.

Fixes #238